+++ src/main/java/io/nats/client/package-info.java
@@ -12,8 +12,8 @@
 // limitations under the License.
 
 /**
- * The Java Nats client is encapsulated into this single package {@code io.nats.client}. Applications
- * will start from the {@link io.nats.client.Nats Nats} class to connect to the gnatsd server. Once connected, they can
+ * The Java NATS client is encapsulated into this single package {@code io.nats.client}. Applications
+ * will start from the {@link io.nats.client.Nats Nats} class to connect to the NATS server. Once connected, they can
  * use the {@link io.nats.client.Connection Connection} object to publish messages or create subscriptions.
  * 
  * @since 2.0.0